Fix for alpha sorting in Chrome with mix-blend-mode lighten
mix-blend-mode: lighten in chrome does not work correctly (it incorrectly darkens the background) unless there is a top level div(i.e. direct child of the div that has the mix-blend-mode property) that will show the background through it. This commit adds a dummy title, which is hidden, that would show the background through it. This allows the actual title and pfp background to properly render the background beneath them.